move rai_interfaces to rai_core
